+++ TinyHttp/Request/MultipartFormContent/BytesMultipartData.cs
@@ -1,0 +1,13 @@
+﻿namespace Tiny.Http
+{
+    internal class BytesMultipartData : MultipartData
+    {
+        public BytesMultipartData(byte[] data, string name, string fileName, string contentType)
+            : base(name, fileName, contentType)
+        {
+            Data = data;
+        }
+
+        public byte[] Data { get; }
+    }
+}