+++ TinyHttp/Request/MultipartFormContent/FileMultipartData.cs
@@ -1,0 +1,15 @@
+﻿using System.IO;
+
+namespace Tiny.Http
+{
+    internal class FileMultipartData : MultipartData, ITinyContent
+    {
+        public FileMultipartData(FileInfo data, string name, string fileName, string contentType)
+            : base(name, fileName, contentType)
+        {
+            Data = data;
+        }
+
+        public FileInfo Data { get; set; }
+    }
+}